+++ src/vmm/src/vstate/vcpu/x86_64.rs
@@ -439,7 +439,7 @@ impl KvmVcpu {
         // per million beacuse it is common for TSC frequency
         // to differ due to calibration at boot time.
         let diff = (i64::from(self.get_tsc_khz()?) - i64::from(state_tsc_freq)).abs();
-        Ok(diff > (f64::from(state_tsc_freq) * TSC_KHZ_TOL).round() as i64)
+        Ok(diff as f64 > (f64::from(state_tsc_freq) * TSC_KHZ_TOL).round())
     }
 
     /// Scale the TSC frequency of this vCPU to the one provided as a parameter.
